+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/persistence/objects/Rating.java
@@ -1,0 +1,14 @@
+package at.fhjoanneum.ippr.gateway.security.persistence.objects;
+
+public interface Rating {
+
+    Long getRatingId();
+
+    Integer getRating();
+
+    void setRating(Integer rating);
+
+    String getComment();
+
+    void setComment(String comment);
+}