+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/persistence/objects/User.java
@@ -1,5 +1,8 @@
 package at.fhjoanneum.ippr.gateway.security.persistence.objects;
 
+import at.fhjoanneum.ippr.gateway.security.persistence.entities.OrganizationImpl;
+
+import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -18,6 +21,22 @@ public interface User {
   void setLastname(String lastname);
 
   String getUsername();
+
+  void setUsername(String username);
+
+  String getEmail();
+
+  void setEmail(String email);
+
+  String getPassword();
+
+  void setPassword(String password);
+
+  OrganizationImpl getOrganization();
+
+  void setOrganization(OrganizationImpl organizationId);
+
+  void setCreatedAt(Date createdAt);
 
   Set<Role> getRoles();