+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/repositories/RBACRepositoryImpl.java
@@ -3,6 +3,8 @@ package at.fhjoanneum.ippr.gateway.security.repositories;
 import java.util.List;
 import java.util.Optional;
 
+import at.fhjoanneum.ippr.gateway.security.persistence.entities.OrganizationImpl;
+import at.fhjoanneum.ippr.gateway.security.persistence.objects.Organization;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -19,6 +21,8 @@ import at.fhjoanneum.ippr.gateway.security.persistence.objects.Role;
 import at.fhjoanneum.ippr.gateway.security.persistence.objects.Rule;
 import at.fhjoanneum.ippr.gateway.security.persistence.objects.User;
 
+import static org.bouncycastle.asn1.x509.X509ObjectIdentifiers.organization;
+
 @Repository
 public class RBACRepositoryImpl implements RBACRepository {
 
@@ -30,6 +34,9 @@ public class RBACRepositoryImpl implements RBACRepository {
 
   @Autowired
   private RuleRepository ruleRepository;
+
+  @Autowired
+  private OrganizationRepository organizationRepository;
 
   @Override
   public User saveUser(final User user) {
@@ -44,6 +51,11 @@ public class RBACRepositoryImpl implements RBACRepository {
   @Override
   public Rule saveRule(final Rule rule) {
     return ruleRepository.save((RuleImpl) rule);
+  }
+
+  @Override
+  public Organization saveOrganization(Organization organization) {
+    return organizationRepository.save((OrganizationImpl) organization);
   }
 
   @Override
@@ -91,7 +103,12 @@ public class RBACRepositoryImpl implements RBACRepository {
     return Lists.newArrayList(ruleRepository.findAll());
   }
 
-  @Repository
+  @Override
+  public Optional<Organization> getOrganizationBySystemId(String systemId) {
+    return Optional.ofNullable(organizationRepository.findBySystemId(systemId));
+  }
+
+    @Repository
   interface UserRepository extends PagingAndSortingRepository<UserImpl, Long> {
 
     @Query(value = "SELECT * FROM user WHERE system_id = :systemId", nativeQuery = true)
@@ -129,5 +146,12 @@ public class RBACRepositoryImpl implements RBACRepository {
 
     @Query(value = "SELECT * FROM rule WHERE system_id = :systemId", nativeQuery = true)
     RuleImpl findBySystemId(@Param("systemId") String systemId);
+  }
+
+  @Repository
+  interface OrganizationRepository extends CrudRepository<OrganizationImpl, Long> {
+
+    @Query(value = "SELECT * FROM organization WHERE system_id = :systemId", nativeQuery = true)
+    OrganizationImpl findBySystemId(@Param("systemId") String systemId);
   }
 }